improvement(zim-to-website): add a stub load.php to reduce console errors
Places a stub where the mediawiki load.php endpoint would be (returns a file with only a comment).
We lose the js that would be loaded in, so we may want to explore statically adding the files that
would be loaded. For now we are just trying to reduce console errors.