fix: use local landing page
In the past we had to download landing from original revision,
but since 2019 ZIM landings improved and can be used as-is.

This switches /wiki/index.html to use "Main Page" from ZIM file
only with minor adjustments (paths, title)

Old logic is commented out in case we need it for ad-hoc fixups
in some language in the future.